This time, I completed the last page and function. Users can input their own body data to calculate BMI, and can set their own goals. Then they can save it. Once the saving is successful, the ideal nutritional intake for achieving this goal will appear on the home page. Moreover, I have optimized the input and calculation methods for food. Users can input the grams of the food they have eaten for precise calculation.

+++ Models/FoodItem.cs
@@ -7,13 +7,9 @@ public class FoodItem
     // Name of the food
     public string Name { get; set; } = "";
     // Calories value
-    public int Calories { get; set; }
-    // Protein amount
-    public int Protein { get; set; }
-    // Carbohydrates amount
-    public int Carbs { get; set; }
-    // Fat amount
-    public int Fat { get; set; }
-    // Water content
-    public int Water { get; set; }
+    public double Calories { get; set; }
+    public double Protein { get; set; }
+    public double Carbs { get; set; }
+    public double Fat { get; set; }
+    public double Water { get; set; }
 }
